Rename file to match changed name and logic
